+++ CHANGELOG.md
@@ -1,3 +1,20 @@
+ORM M2 - 0.6.1.24
+------
+- 0006289: Dans le IF pour savoir si l'événement est l'événement de l'organisateur ajouter le test s'il est externe
+- 0006291: Permettre un type booleanLdap sur une entrée multivaluée
+- 0006292: Ajouter une propriété is_mailbox dans l'objet User
+- 0006288: Lorsqu'on recherche si l'organisateur est externe, valider l'objectClass mineqMelBoite
+
+ORM M2 - 0.6.1.23
+------
+- Mise à jour du schéma de base de données initial
+
+ORM M2 - 0.6.1.22
+------
+- 0006244: Problème dans le calcul de la date de fin à partir du count pour une modification
+- 0006245: Optimisation du calcul de la date de fin à partir du count de la recurrence
+- 0006232: [En attente] Gérer les catégories des espaces de travail du BNum
+
 ORM M2 - 0.6.1.21
 ------
 - Mise à jour de l'API MI pour les shares et vacation